# grpc/golang/api/helloworld/stock_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// protoc-gen-go-http v2.2.2

package helloworld

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

type StockHTTPServer interface {
	SayHello(context.Context, *StockRequest) (*StockReply, error)
}

func RegisterStockHTTPServer(s *http.Server, srv StockHTTPServer) {
	r := s.Route("/")
	r.POST("/stock/stock_info", _Stock_SayHello0_HTTP_Handler(srv))
	r.GET("/stock/{name}", _Stock_SayHello1_HTTP_Handler(srv))
}

func _Stock_SayHello0_HTTP_Handler(srv StockHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in StockRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, "/helloworld.Stock/SayHello")
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SayHello(ctx, req.(*StockRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*StockReply)
		return ctx.Result(200, reply)
	}
}

func _Stock_SayHello1_HTTP_Handler(srv StockHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in StockRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, "/helloworld.Stock/SayHello")
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SayHello(ctx, req.(*StockRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*StockReply)
		return ctx.Result(200, reply)
	}
}

type StockHTTPClient interface {
	SayHello(ctx context.Context, req *StockRequest, opts ...http.CallOption) (rsp *StockReply, err error)
}

type StockHTTPClientImpl struct {
	cc *http.Client
}

func NewStockHTTPClient(client *http.Client) StockHTTPClient {
	return &StockHTTPClientImpl{client}
}

func (c *StockHTTPClientImpl) SayHello(ctx context.Context, in *StockRequest, opts ...http.CallOption) (*StockReply, error) {
	var out StockReply
	pattern := "/stock/{name}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation("/helloworld.Stock/SayHello"))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}